Refactor projection function to use helper functions and update SimulParams documentation to clarify default distortion_center value
