+++ calibration/simulator/camera.py
@@ -1,40 +1,34 @@
-from dataclasses import dataclass
+from dataclasses import dataclass, field
 
 import numpy as np
 
 
+# TODO: use cached_property
 @dataclass
 class Camera:
     focal_length: float = 35.0
-    sensor_size: tuple = (36.0, 24.0)
-    resolution: tuple = (1200, 800)
+    sensor_size: np.ndarray = field(default_factory=lambda: np.array([36.0, 24.0]))
+    resolution: np.ndarray = field(default_factory=lambda: np.array([1200, 800]))
     skew: float = 0.0
 
     @property
-    def pixel_size(self) -> tuple:
-        return (
-            self.sensor_size[0] / self.resolution[0],
-            self.sensor_size[1] / self.resolution[1],
-        )
+    def pixel_size(self) -> np.ndarray:
+        return self.sensor_size / self.resolution
 
     @property
-    def fx(self) -> float:
-        return self.focal_length / self.pixel_size[0]
+    def focal_lenght_pixels(self) -> np.ndarray:
+        return self.focal_length / self.pixel_size
 
     @property
-    def fy(self) -> float:
-        return self.focal_length / self.pixel_size[1]
-
-    @property
-    def image_center(self) -> tuple[float, float]:
-        return self.resolution[0] / 2, self.resolution[1] / 2
+    def image_center(self) -> np.ndarray:
+        return self.resolution / 2
 
     @property
     def intrinsic_matrix(self) -> np.ndarray:
-        return np.array(
-            [
-                [self.fx, self.skew, self.image_center[0]],
-                [0, self.fy, self.image_center[1]],
-                [0, 0, 1],
-            ]
-        )
+        fx, fy = self.focal_lenght_pixels
+        cx, cy = self.image_center
+        return np.array([[fx, self.skew, cx], [0, fy, cy], [0, 0, 1]])
+
+    @property
+    def principal_point(self) -> np.ndarray:
+        return self.image_center * self.pixel_size
